requirements.txt
alpaca-py==0.10.0
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.12.2
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.5
django-allauth==0.52.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
frozendict==2.3.8
gitdb==4.0.10
GitPython==3.1.30
graphene==3.3
graphene-django==3.1.5
graphene-django-crud==2.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
html5lib==1.1
idna==3.4
lxml==4.9.2
msgpack==1.0.5
multitasking==0.0.11
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
pandas==2.1.0
pathspec==0.10.3
platformdirs==2.6.2
promise==2.3
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
smmap==5.0.0
soupsieve==2.5
sqlparse==0.4.3
sseclient-py==1.8.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
webencodings==0.5.1
websockets==10.4
yfinance==0.2.4
